major refactor of utility functions
