+++ src/lib/models/DiscoverGraph.ts
@@ -30,9 +30,7 @@ export class DiscoverGraph {
     }
 
     public addEdge(edge: DiscoverEdge) {
-        //console.log("adding ", edge);
         this._edges.add(edge);
-        //console.log("edges ", this._edges);
     }
 
     public pushOrigin(origin: pbReadInfoResponse, init: boolean) {
@@ -94,7 +92,8 @@ export class DiscoverGraph {
                     isContributor: true,
                     isContribution: false,
                     from: repo_id,
-                    to: contributor.login
+                    to: contributor.login,
+                    num_contributions: contributor.contributions
                 }
 
                 // add edge to graph
@@ -126,7 +125,7 @@ export class DiscoverGraph {
                     isContributor: false,
                     isContribution: true,
                     from: login,
-                    to: contribution.nameWithOwner
+                    to: contribution.nameWithOwner,
                 }
 
                 // add edge to graph
@@ -141,20 +140,6 @@ export class DiscoverGraph {
             edges: [...this._edges.values()]
         };
     }
-
-    private ensureRepositoryNodeExists(origin: DiscoverNode) {
-        if(!origin.id) {
-            const baseOrigin: DiscoverNode = {
-                id: "Repository Not Found.",
-                label: "Repository Not Found.",
-                isOrigin: true,
-                isRepository: true,
-                avatar_url: defaultAvatarUrl,
-            };
-
-            this._nodes.set("Repository Not Found.", baseOrigin);
-        }
-    }
 }
 
 export interface DiscoverNode extends vis.Node {
@@ -168,6 +153,7 @@ export interface DiscoverNode extends vis.Node {
 export interface DiscoverEdge extends vis.Edge {
   isContributor?: boolean;
   isContribution?: boolean;
+  num_contributions?: number;
 }
 
 export interface NodeEdgeArray {
