+++ tsconfig.json
@@ -8,6 +8,5 @@
 		"resolveJsonModule": true,
 		"skipLibCheck": true,
 		"sourceMap": true,
-		"strict": true
 	},
 }
